+++ Structurizr.Core/Model/Tags.cs
@@ -14,6 +14,7 @@
         public const string Synchronous = "Synchronous";
         public const string Asynchronous = "Asynchronous";
 
+        public const string DeploymentNode = "Deployment Node";
         public const string ContainerInstance = "Container Instance";
         
     }